requirements.txt
openai>=1.30.0
httpx>=0.27.0
pydantic>=2.7.0
pyyaml>=6.0.1
pytest>=8.2.0
pytest-asyncio>=0.23.0
